requirements.txt
Django==5.0.3
djangorestframework==3.15.1
drf-spectacular==0.27.1
django-import-export==3.3.7

